fix: renamed (added "Param") query is bad, queryParam good.
